refactor: unify landlord messages and student messages page
